feat(speakers): use id instead of position in firebase
Some of the devfests use cfp.io which allow us to extract data in json to integrate it with the hoverboard project (with some `jq` transformations). In this system, the speakers and talks already have ids, so I made modification on the hoverboard project to handle speakers by id instead of position in array. This is much more simple to use and allow us to remove a speaker when we want (even if he wasn't the last of speakers).